\name{CORREP-package}
\alias{CORREP-package}
\alias{CORREP}
\docType{package}
\title{
Multivariate Correlation Estimator from Replicated Data
}
\description{
This package implements multivariate correlation estimator for data with 
equivalent number of replicates as well as inequivalent number of replicates. 
For small sample data, this package provides permutation test for test if
the true correlation vanishes, and for large sample data, this package provides 
Likelihood Ratio Test (LRT) to test if the true correlation vanishes. A function 
to calculate Bootstrap confidence interval was also included in the package.  
}
\details{
\tabular{ll}{
Package: \tab CORREP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-03-19\cr
License: \tab GPL 2.0 or newer\cr
}
}
\author{
Dongxiao Zhu and Youjuan Li
Maintainer: Dongxiao Zhu <doz@stowers-institute.org>
}
\references{
Zhu, D and Li Y. 2007. Multivariate Correlation Estimator for Inferring Functional Relationships
from Replicated 'OMICS' data. Submitted.
}
\keyword{multivariate}
\keyword{cluster}
\keyword{models}
\keyword{htest}
\examples{
}
