\name{CORREP-package}
\alias{CORREP-package}
\alias{CORREP}
\docType{package}
\title{
Implment multivariate correlation estimator from replicated data
CORREP
}
\description{
}
\details{
\tabular{ll}{
Package: \tab CORREP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-03-19\cr
License: \tab What license is it under?\cr
}
}
\author{
Dongxiao Zhu and Youjuan Li
Maintainer: Dongxiao Zhu <doz@stowers-institute.org>
}
\references{
Zhu, D and Li Y. 2007. Multivariate Correlation Estimator for Inferring Functional Relationships
from Replicated 'OMICS' data. Submitted.
}
\keyword{multivariate}
\keyword{cluster}
\keyword{models}
\keyword{htest}
\examples{
}
